+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+
+## 3.6.1
+
+**2026-04-26**
+
 ### Fixed
 
 - Broken Markdown navigation on QR views.